+++ x/fungible/client/cli/query_code_hash.go
@@ -1,0 +1,36 @@
+package cli
+
+import (
+	"context"
+
+	"github.com/cosmos/cosmos-sdk/client"
+	"github.com/cosmos/cosmos-sdk/client/flags"
+	"github.com/spf13/cobra"
+	"github.com/zeta-chain/zetacore/x/fungible/types"
+)
+
+func CmdQueryCodeHash() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "code-hash [address]",
+		Short: "shows the code hash of an account",
+		Args:  cobra.ExactArgs(1),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx := client.GetClientContextFromCmd(cmd)
+
+			queryClient := types.NewQueryClient(clientCtx)
+
+			res, err := queryClient.CodeHash(context.Background(), &types.QueryCodeHashRequest{
+				Address: args[0],
+			})
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(res)
+		},
+	}
+
+	flags.AddQueryFlagsToCmd(cmd)
+
+	return cmd
+}