+++ x/fungible/client/cli/tx_update_contract_bytecode.go
@@ -1,0 +1,38 @@
+package cli
+
+import (
+	"github.com/cosmos/cosmos-sdk/client"
+	"github.com/cosmos/cosmos-sdk/client/flags"
+	"github.com/cosmos/cosmos-sdk/client/tx"
+	"github.com/spf13/cobra"
+	"github.com/zeta-chain/zetacore/x/fungible/types"
+)
+
+func CndUpdateContractBytecode() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "update-contract-bytecode [contract-address] [new-code-hash]",
+		Short: "Broadcast message UpdateContractBytecode",
+		Args:  cobra.ExactArgs(2),
+		RunE: func(cmd *cobra.Command, args []string) (err error) {
+
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			msg := types.NewMsgUpdateContractBytecode(
+				clientCtx.GetFromAddress().String(),
+				args[0],
+				args[1],
+			)
+			if err := msg.ValidateBasic(); err != nil {
+				return err
+			}
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}